requirements.txt
# PDF处理
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# 翻译API
openai>=1.0.0
google-cloud-translate>=3.12.0
httpx>=0.25.0

# 工具
pyyaml>=6.0
click>=8.1.0
tqdm>=4.66.0
Pillow>=10.0.0

# 可选: 本地LLM支持
# vllm>=0.2.0
